consolidate types for entities folder
