+++ src/entities/actionpoint-entity.ts
@@ -1,17 +1,14 @@
 import { PropertyNames } from 'helpers';
 import EntityConfig from 'entities';
-
-export interface ActionPointEntity {
-    id: number;
-    reference_number: string;
-    description: string;
-    status: string;
-}
+import { ActionPointEntity } from './types';
 
 
 export default class ActionPointConfig extends EntityConfig<ActionPointEntity> {
     public get displayProperties(): PropertyNames<ActionPointEntity>[] {
         return ['reference_number', 'description'];
+    }
+    public get title() {
+        return 'Action Points';
     }
 }
