+++ src/entities/indicator-entity.ts
@@ -1,16 +1,14 @@
 import EntityConfig from 'entities';
 import { PropertyNames } from 'helpers';
-
-export interface IndicatorEntity {
-    id: number;
-    title: string;
-    section: number;
-}
+import { IndicatorEntity } from './types';
 
 
 export default class IndicatorConfig extends EntityConfig<IndicatorEntity> {
     public get displayProperties(): PropertyNames<IndicatorEntity>[] {
         return ['title', 'id'];
+    }
+    public get title() {
+        return 'Applied Indicators';
     }
 }
