+++ src/entities/tpmactivity-entity.ts
@@ -1,17 +1,16 @@
 import EntityConfig from 'entities';
 import { PropertyNames } from 'helpers';
-
-export interface TPMActivityEntity {
-    id: number;
-    reference_number: string;
-    status: string;
-}
+import { TPMActivityEntity } from './types';
 
 
 export default class TPMActivityConfig extends EntityConfig<TPMActivityEntity> {
 
     public get displayProperties(): PropertyNames<TPMActivityEntity>[] {
         return ['reference_number', 'id'];
+    }
+
+    public get title() {
+        return 'TPM Activities';
     }
 
 }
